requirements.txt
gensim==3.8.3
toml==0.10.1
matplotlib==3.3.2
nltk==3.5
pandas==1.1.1
numpy==1.19.1
play-scraper==0.6.0
glove-python==0.1.0
langdetect~=1.0.8
requests~=2.25.0